Retry Button + Minor Adjustments
Retry button added for pause menu in all 5 levels.
Made collision boxes of parked cars on level 2 more accurate/tight.
Moved the fog at the start of level 5 back so it's not covering the car and drew some extra tiles at the start to make the sea less visible when the fog is further back.